Package and naming refactoring
