+++ subs-samples-agent/src/main/java/uk/ac/ebi/subs/SamplesAgentApp.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.subs.samplesagent;
+package uk.ac.ebi.subs;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
