+++ subs-samples-agent/src/main/java/uk/ac/ebi/subs/samplesagent/SamplesListenerConfiguration.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.subs.samplesagent.messaging;
+package uk.ac.ebi.subs.samplesagent;
 
 import org.springframework.amqp.rabbit.annotation.RabbitListenerConfigurer;
 import org.springframework.amqp.rabbit.listener.RabbitListenerEndpointRegistrar;
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @Configuration
-public class ReceiverConfiguration implements RabbitListenerConfigurer {
+public class SamplesListenerConfiguration implements RabbitListenerConfigurer {
 
     @Override
     public void configureRabbitListeners(RabbitListenerEndpointRegistrar registrar) {
